+++ RELEASENOTES.md
@@ -1,3 +1,7 @@
+## [0.6.0]
+
+- Some technical changes.
+
 ## [0.5.0]
 
 ### FIxed